final report, refactor, bug fix

+++ source/ajax/unit-lector-delete.php
@@ -8,11 +8,16 @@
     // get parameters
 	$unit_id = post_escaped("unitID");
 	$lector_id = post_escaped("lectorID");
+
+    if (!require_user_editor($mysqli, $unit_id, false)) {
+        echo "Nemate opravnenie.";
+        die();
+    }
 	
 	// query
 	$query = "DELETE FROM unit_account WHERE unit_id=$unit_id AND account_id = $lector_id";
 		if (!($result = $mysqli->query($query))) {
-			echo 'nevyslo to';
+			echo 'ERROR';
 		} else {
 			echo 'true';
 		}
